refactor(#114): split topic controller/service by resource
God-files forced every academic-content change through one module; split by @ApiTags so each resource has one reason to change, without touching routes or API contracts.
