+++ docs/Database Schema.md
@@ -344,7 +344,7 @@ Tài liệu này được tổng hợp trực tiếp từ Prisma schema tại `a
   - `POST /courses` — cần `name`; `default_duration_days` (để trống = vô hạn) và `sort_order` tuỳ chọn; `id` tự sinh.
   - `PATCH /courses/:id` — cập nhật `name`/`default_duration_days`/`sort_order`/`is_active` khi field được truyền; truyền `default_duration_days: null` để chuyển về vô hạn.
   - `DELETE /courses/:id` — `400` nếu còn lớp đang dùng khoá học này (`classes.count > 0`); thông báo hướng dẫn chuyển lớp sang khoá học khác hoặc set `is_active=false` thay vì xoá cứng.
-- Guard phân quyền nội dung khoá (reusable `CourseAccessService`, dùng lại cho mọi ticket nội dung khoá về sau): admin đầy đủ / trợ lí / trưởng giáo án quản lý được mọi khoá; thành viên `lesson_plan` chỉ thấy/sửa khoá mình được gán (qua `course_lesson_plan_members`); gia sư đang dạy lớp thuộc khoá X **không** vì thế mà sửa được nội dung cấp khoá của X. `TopicService` gọi `assertCanManageCourse` trước mọi ghi Chapter / Topic nhánh khoá / Lecture / Thư viện đề thi và trước GET câu hỏi/quiz trả đáp án cấp khoá; `assertCanWriteCourseQuestions` chỉ áp dụng khi ghi **ngân hàng câu hỏi**.
+- Guard phân quyền nội dung khoá (reusable `CourseAccessService`, dùng lại cho mọi ticket nội dung khoá về sau): admin đầy đủ / trợ lí / trưởng giáo án quản lý được mọi khoá; thành viên `lesson_plan` chỉ thấy/sửa khoá mình được gán (qua `course_lesson_plan_members`); gia sư đang dạy lớp thuộc khoá X **không** vì thế mà sửa được nội dung cấp khoá của X. Các service resource trong `TopicModule` gọi `assertCanManageCourse` (qua `TopicSupportService`) trước mọi ghi Chapter / Topic nhánh khoá / Lecture / Thư viện đề thi và trước GET câu hỏi/quiz trả đáp án cấp khoá; `assertCanWriteCourseQuestions` chỉ áp dụng khi ghi **ngân hàng câu hỏi**.
 - Seed dữ liệu ban đầu gồm các mã cũ (`vip`, `basic`, `advance`, `hardcore`) cộng 3 mã mới: `thpt_basic` (THPT BASIC), `thpt_advanced` (THPT ADVANCED), `thpt_luyen_de` (THPT Luyện Đề).
 - Khi tạo lớp không truyền `course_id`, `ClassService.resolveDefaultCourseId` fallback về khoá học `is_active=true` có `sort_order` nhỏ nhất (tie-break theo `name`) — không còn hardcode `code='basic'` để tránh vỡ khi admin đổi/xoá phân loại mặc định cũ.
