+++ docs/pages/admin.md
@@ -31,7 +31,7 @@
 - **Cài đặt khoá** (`/admin/classes/courses/:id`, Màn 08): trang cấu hình một Khoá học — đổi tên/thời hạn mặc định/thứ tự/bật tắt (cùng `CourseFormPopup` + `Switch`), xoá khoá và xoá mức độ khó qua `ConfirmDialog` (không `window.confirm`), và quản lý hai phần: **Thang mức độ khó** (thêm/đổi tên/đổi thứ tự bằng nút ↑↓/tắt/bật/xoá; danh sách gọi `GET /courses/:id/difficulty-levels` qua `useCourseDifficultyLevels(courseId, true)` gồm inactive; thao tác qua nested endpoints; invalidate `courseKeys.difficultyLevelsPrefix(courseId)` để picker active-only cùng khoá cũng refetch; dùng `runBackgroundSave` + Sonner) và **Đội giáo án** (danh sách member + nút Gỡ; picker `UpgradedSelect` searchable tìm nhân sự active có role `lesson_plan` qua `GET /courses/lesson-plan-staff`, Thêm/`PUT /courses/:id/lesson-plan-members` thay toàn bộ — nút Thêm/Gỡ disable khi mutation `isPending`, nhãn **Đang lưu…**). Khi sửa một khoá đã tồn tại, popup thời hạn mặc định hiển thị banner cảnh báo dạng `Alert variant="warning"`: **"Chỉ áp dụng cho lớp tạo mới — Lớp đang chạy giữ nguyên hạn đã đặt. Đổi hạn từng lớp ở trang lớp."** `lesson_plan_head` không cần gán (luôn có toàn quyền mọi khoá) nên không xuất hiện trong picker để thêm; nếu có mặt trong danh sách member thì hiện pill **Mặc định** và không có nút Gỡ. Chỉ admin đầy đủ và trợ lí thao tác; người gán đội giáo án giới hạn ở admin/`lesson_plan_head`/`assistant` (BE). **Cây tri thức** (`KnowledgeTreeCard`): lúc `isLoading` hiện shadcn `Skeleton` (không chữ "Đang tải..."); dialog **Sửa bài học** seed lại `lecture_quizzes` mỗi lần mở (kể cả cùng `lecture.id`); gỡ quiz thật sự thì `ConfirmDialog` trước khi unlink. Ngân hàng quiz trong dialog gọi `getQuestions({ courseId })` (query key `questionKeys.list({ courseId })`) — không lấy câu hỏi khoá khác. Shape cây dùng `KnowledgeTreeNode` / `KnowledgeTreeTopicNode` trong `apps/web/dtos/topic.dto.ts`. Kéo-thả Chủ đề / Chuyên đề / Bài học cập nhật cache ngay, revert khi API lỗi; drag handle `touch-none` + `{...attributes}` (Space/mũi tên); kéo chuyên đề sang chủ đề khác toast `"Không thể chuyển chương ở đây"` (không chuyển chương trên UI này).
 - **Thư viện đề thi** (`/admin/exam-library`, Màn 03): trang quản lý đề thi (chuyên đề luyện tập `Topic.kind = practice`) ở cấp khoá học, `chapterId = null`, `courseId` có giá trị. Đề thi trong thư viện dùng chung cho mọi lớp thuộc cùng khoá; gia sư thêm đề vào lớp qua tab Nội dung lớp (`createClassContent` với `topicId`). Page gồm: chọn khoá học (`UpgradedSelect`), tìm kiếm đề (debounce 300ms rồi mới gọi API), tạo/sửa tên/xóa đề thi (`ConfirmDialog`, không `window.confirm`), và mở rộng để quản lý câu hỏi liên kết (`PracticeTopicQuestionsCard` dùng chung với ticket #55). **Sửa tên inline:** Enter lưu, Escape huỷ (không lưu — chặn `onBlur` ghi đè), rời ô thì lưu; chú thích dưới ô nhập. CRUD đề thi qua `GET/POST/PATCH/DELETE /course/:courseId/exam-library` và `POST /course/:courseId/exam-library/reorder`. Câu hỏi liên kết tái sử dụng module #55: `GET/POST/PATCH/DELETE /topics/:topicId/questions` (cùng `summary`, `is-assigned`, `reorder`). Mở cho admin đầy đủ, trợ lí, `lesson_plan` (chỉ khoá được gán), `lesson_plan_head` (mọi khoá). Ghi/đọc đáp án cấp khoá gọi `CourseAccessService.assertCanManageCourse` — HTTP 403 nếu không thuộc đội giáo án. Frontend dùng `examLibraryKeys.course(courseId)` khi invalidate (không `examLibraryKeys.all`) và `practiceTopicQuestionKeys` cho câu hỏi. List lúc `isLoading` hiện shadcn `Skeleton`; empty-state chỉ sau khi load xong. Nút tạo/sửa/xóa disable khi mutation `isPending` (nhãn **Đang lưu…**). `PracticeTopicQuestionsCard` cùng quy tắc skeleton vs empty (kể cả dialog thêm từ ngân hàng); ô tìm trong dialog debounce 300ms.
 - **Ngân hàng câu hỏi** (`/admin/question-bank`): CRUD câu hỏi cấp khoá. DTO `Course` / `Chapter` / `CourseDifficultyLevel` import từ `apps/web/dtos/` (không khai báo cục bộ trên page). List lúc `isLoading` hiện shadcn `Skeleton`. Nút **Nhập từ AI** disable khi chưa chọn khoá, tooltip native `title="Chọn khoá học trước"`; đã chọn thì mở `AiImportModal`. `chapters` / `difficulty-levels` lấy qua `useCourseChapters` / `useCourseDifficultyLevels` (cùng hook với BankPicker, `PracticeTopicQuestionsCard`, `AiImportModal`, `ClassPracticeQuestionComposer`). Mutation create/update/delete invalidate `questionKeys.course(courseId)` (+ list không lọc `["question","list"]`) — không `questionKeys.all` khi đã biết khoá.
-- **Ma trận quyền nội dung cấp khoá (dạy lớp ≠ soạn giáo án):** `TopicService` bọc `assertCanManageCourse` trên mọi ghi Chapter / Topic nhánh khoá / Lecture / Thư viện đề thi và trên GET trả `correctIndex`/`explanation`/`answerGuide` cấp khoá. Role decorator (`StaffRole.teacher`) **không** đủ — phải là admin đầy đủ, trợ lí, trưởng giáo án, hoặc `lesson_plan` được gán khoá (`course_lesson_plan_members`).
+- **Ma trận quyền nội dung cấp khoá (dạy lớp ≠ soạn giáo án):** các service resource trong `TopicModule` (`CourseChapterService`, `CourseTopicService`, `LectureService`, `ExamLibraryService`, `PracticeQuestionLinkService`) gọi `CourseAccessService.assertCanManageCourse` (qua `TopicSupportService.assertCanManageCourseContent`) trên mọi ghi Chapter / Topic nhánh khoá / Lecture / Thư viện đề thi và trên GET trả `correctIndex`/`explanation`/`answerGuide` cấp khoá. Role decorator (`StaffRole.teacher`) **không** đủ — phải là admin đầy đủ, trợ lí, trưởng giáo án, hoặc `lesson_plan` được gán khoá (`course_lesson_plan_members`).
 
   | Hành động | admin / `assistant` / `lesson_plan_head` | `lesson_plan` gán khoá | `teacher` dạy lớp thuộc khoá, không trong đội giáo án |
   | --- | --- | --- | --- |
